feature/BB:convert-id-to-int64
Convert type id for all project to int64.

+++ bankBackend/internal/core/currency.go
@@ -1,14 +1,14 @@
 ﻿package core
 
 type Currency struct {
-	Id uint64 `json:"id" db:"id"`
+	Id int64 `json:"id" db:"id"`
 
 	Name string `json:"name" db:"name"`
 
 	Symbol  rune   `json:"symbol" db:"symbol"`
 	IsoCode string `json:"iso_code" db:"iso_code"`
 
-	MinorUnits int `json:"minor_units" db:"minor_units"`
+	MinorUnits int8 `json:"minor_units" db:"minor_units"`
 }
 
 type CurrencyCreateInput struct {
@@ -17,7 +17,7 @@ type CurrencyCreateInput struct {
 	Symbol  rune   `json:"symbol" db:"symbol"`
 	IsoCode string `json:"iso_code" db:"iso_code"`
 
-	MinorUnits int `json:"minor_units" db:"minor_units"`
+	MinorUnits int8 `json:"minor_units" db:"minor_units"`
 }
 
 type CurrencyUpdateInput struct {
